+++ CHANGES.rst
@@ -3,6 +3,11 @@ Changes
 
 In next release ...
 
+- Added ``RenderError`` exception which indicates that an error
+  occurred during the evaluation of an expression.
+
+- Clean up ``TemplateError`` exception implementation.
+
 
 2.20 (2015-01-12)
 -----------------